+++ Mousecape/Mousecape/src/controllers/MCLibraryController.h
@@ -13,13 +13,14 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface MCLibraryController : NSObject
 @property (readonly, weak, nullable) MCCursorLibrary *appliedCape;
-@property (nonatomic, readonly) NSUndoManager *undoManager;
 @property (readonly, copy) NSURL *libraryURL;
 
 - (instancetype)initWithURL:(NSURL *)url;
 
 - (nullable NSError *)importCapeAtURL:(NSURL *)url;
+- (nullable NSError *)importCapeAtURL:(NSURL *)url skipValidation:(BOOL)skipValidation;
 - (nullable NSError *)importCape:(MCCursorLibrary *)cape;
+- (nullable NSError *)importCape:(MCCursorLibrary *)cape skipValidation:(BOOL)skipValidation;
 
 - (void)addCape:(MCCursorLibrary *)cape;
 - (void)removeCape:(MCCursorLibrary *)cape;
@@ -28,9 +29,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)restoreCape;
 
 - (NSURL *)URLForCape:(MCCursorLibrary *)cape;
-
-- (NSSet *)capesWithIdentifier:(NSString *)identifier;
-- (BOOL)dumpCursorsWithProgressBlock:(BOOL (^)(NSUInteger current, NSUInteger total))block;
 
 @end